Nested friendslist component to be side by side with chatroom
