+++ small-talk/styles/globals.css
@@ -1,3 +1,15 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+
+/* custom.css */
+.chat-container {
+    @apply flex justify-between items-start;
+}
+
+@media (max-width: 600px) {
+    .chat-container {
+        @apply flex-col;
+    }
+}
